TFTDataSet/requirements.txt
asgiref==3.2.10
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.1.2
django-environ==0.4.5
djangorestframework==3.12.1
idna==3.3
psycopg2==2.8.6
pytz==2020.1
requests==2.27.1
riotwatcher==3.2.2
sqlparse==0.4.1
urllib3==1.26.9
